Refactor store-scoped operation handling in tests.
Extract `StoreScopedOperationTrait` and `ResourceOperationTrait` to consolidate common logic. Replaced inheritance from `AbstractTestStoreScopedCreateOperation` with traits for better flexibility and modularity. Simplified and clarified the test structure for store-scoped resource operations.
