+++ tests/Integration/Resource/StoreScopedOperationTrait.php
@@ -4,8 +4,10 @@ declare(strict_types=1);
 
 namespace Tests\Integration\Resource;
 
-abstract class AbstractTestStoreScopedCreateOperation extends AbstractTestCreateOperation
+trait StoreScopedOperationTrait
 {
+    use ResourceOperationTrait;
+
     protected string $storeId = '9d024f41-3faf-4eef-9d2d-a7e506b81afb';
 
     abstract protected function getStoreScopedResourcePath(): string;
@@ -15,9 +17,6 @@ abstract class AbstractTestStoreScopedCreateOperation extends AbstractTestCreate
         return $this->normalizePath('stores', $this->storeId, $this->getStoreScopedResourcePath());
     }
 
-    /**
-     * Helper method to construct store-scoped URIs.
-     */
     protected function buildStoreUri(string ...$segments): string
     {
         return $this->buildUri('stores', $this->storeId, ...$segments);
